Changed int returns to ssize_t.

+++ cartectrl/cartectrl/include/usb/usb_com.h
@@ -15,6 +15,7 @@
 #include <stdint.h>
 #include <vector>
 #include <string>
+#include <sys/types.h>
 
 #if __GXX_EXPERIMENTAL_CXX0X__ == 1 /* require c++11 support */
  #include <array>
@@ -43,7 +44,7 @@ using std::string;
 int init();
 
 // Base trasmit function
-int write(const uint8_t *m, size_t s);
+ssize_t write(const uint8_t *m, size_t s);
 
 // Convenience functions:
 
@@ -53,26 +54,26 @@ int write(const uint8_t *m, size_t s);
 // mind the endianness of both ends of the communication. You might use a POD
 // with no pointer member. Anything else is just plain silly.
 template <typename T>
- inline int write(const T *m, size_t s)
+ inline ssize_t write(const T *m, size_t s)
 {
     return write(reinterpret_cast<const uint8_t *>(m), s*sizeof(T));
 }
 
 // Adaptor: any vector (see comment above for restrictions on type T).
 template <typename T>
- inline int write(const vector<T> &v)
+ inline ssize_t write(const vector<T> &v)
 {
     return write(v.data(), v.size());
 }
 // Adaptor: a string.
-inline int write(const string &s)
+inline ssize_t write(const string &s)
 {
     return write(s.data(), s.length());
 }
 
 // Adaptor: any plain array (see comment above for restrictions on type T).
 template <typename T, size_t N>
- inline int write(const T (&arr)[N])
+ inline ssize_t write(const T (&arr)[N])
 {
     return write(arr, N);
 }
@@ -81,7 +82,7 @@ template <typename T, size_t N>
 
 // Adaptor: any std::array (see comment above for restrictions on type T).
 template <typename T, size_t N>
- inline int write(const array<T, N> &arr)
+ inline ssize_t write(const array<T, N> &arr)
 {
     return write(arr.data(), N);
 }
@@ -89,7 +90,7 @@ template <typename T, size_t N>
 #endif
 
 // Base trasmit function
-int read(uint8_t *m, size_t s);
+ssize_t read(uint8_t *m, size_t s);
 
 // Convenience functions:
 
@@ -97,20 +98,20 @@ int read(uint8_t *m, size_t s);
 // object type in it, or you'll get pretty Bad Results™, ESPECIALLY more that
 // one that the write overload above. The same comment does apply.
 template <typename T>
- inline int read(T *m, size_t s)
+ inline ssize_t read(T *m, size_t s)
 {
     return read(reinterpret_cast<uint8_t *>(m), s*sizeof(T));
 }
 
 // Adaptor: any vector (see comment above for restrictions on type T).
 template <typename T>
- inline int read(vector<T> &v)
+ inline ssize_t read(vector<T> &v)
 {
     // TODO: Do it better!
     return read(v.data(), v.size());
 }
 // Adaptor: a string.
-inline int read(string &s)
+inline ssize_t read(string &s)
 {
     // TODO: Do it better!
     return read(&*s.begin(), s.length());
@@ -118,7 +119,7 @@ inline int read(string &s)
 
 // Adaptor: any plain array (see comment above for restrictions on type T).
 template <typename T, size_t N>
- inline int read(T (&arr)[N])
+ inline ssize_t read(T (&arr)[N])
 {
     return read(arr, N);
 }
@@ -127,7 +128,7 @@ template <typename T, size_t N>
 
 // Adaptor: any std::array (see comment above for restrictions on type T).
 template <typename T, size_t N>
- inline int read(array<T, N> &arr)
+ inline ssize_t read(array<T, N> &arr)
 {
     return read(arr.data(), N);
 }
